[optimize] Type folder structure
